Moved set_quantity to utils
